consolidate sd into a single file
This allows us to run:

    sd command --help

Instead of just:

    sd help command

And makes all of the sdefaults/ thin wrappers around built-in functionality.

This is a breaking change, though, as previously --help would have
been passed through to the underlying script. But it's a step in the direction
I want to go in: over time I want to make sd smarter, and add some argument
parsing based comments. So it will be doing more of this.