complete arguments for scripts
When completion reaches a script, it will continue trying to complete
positional arguments with the standard _files completer. It will also complete
the built-in flags understood by sd, but only if you begin with a hyphen.

It will also complete --new if you enter a script path that does not exist yet.