fewer forks when using sd in function or executable mode
This removes the subshell in the main dispatch function when using sd as a
regular script -- although it's still present when using it as a shell function.

It also execs scripts when it finds them instead of forking to them. The effect
is the same, but there's less nesting if you look at the running process tree.